broadcom-poc/agent-backend/requirements.txt
Flask>=2.0
python-dotenv>=1.0.0
vectara-agentic>=0.1.0 # Replace with the specific version if needed
flask-cors>=3.0.10 
fastapi>=0.100.0
uvicorn>=0.22.0
nest_asyncio>=1.5.6
sendgrid>=6.9.7 # Add SendGrid 